Refactoring
Moving common code to htmlHelper and suggstSupportHelper. Also removing
Yaml support and only leaving in the Docker-compose.yml support.
