+++ dockerfile/dockerfileExtraInfo.ts
@@ -6,6 +6,7 @@
 
 import vscode = require('vscode');
 import {Range} from 'vscode';
+import htmlHelper = require('../helpers/htmlHelper');
 
 interface InstructionToken {
 	startIndex: number;
@@ -36,7 +37,7 @@ export class ExtraInfoSupport implements vscode.Modes.IExtraInfoSupport  {
 		
 		var result: vscode.Modes.IComputeExtraInfoResult = {
 			range: new Range(position.line, textToken.startIndex, position.line, lineEndColumn),
-			htmlContent: simpleMarkDownToHTMLContent(info)
+			htmlContent: htmlHelper.simpleMarkDownToHTMLContent(info)
 		}
 
 		return Promise.resolve(result);
@@ -70,79 +71,6 @@ export class ExtraInfoSupport implements vscode.Modes.IExtraInfoSupport  {
 			instruction: firstWord,
 		}
 	}
-}
-
-// Copied from yaml
-enum TagType {
-	pre,
-	bold,
-	normal
-}
-
-function simpleMarkDownToHTMLContent(source:string): vscode.IHTMLContentElement[] {
-	var r:vscode.IHTMLContentElement[] = [];
-
-	var lastPushedTo:number;
-	var push = (to:number, type:TagType) => {
-		if (lastPushedTo >= to) {
-			return;
-		}
-		var text = source.substring(lastPushedTo, to);
-
-		if (type === TagType.pre) {
-			r.push({
-				tagName: "span",
-				style: "font-family:monospace",
-				className: "token keyword",
-				text: text
-			});
-		} else if (type === TagType.bold) {
-			r.push({
-				tagName: "strong",
-				text: text
-			});
-		} else if (type === TagType.normal) {
-			r.push({
-				tagName: "span",
-				text: text
-			});
-		}
-		lastPushedTo = to;
-	}
-
-	var currentTagType = () => {
-		if (inPre) {
-			return TagType.pre;
-		}
-		if (inBold) {
-			return TagType.bold;
-		}
-		return TagType.normal;
-	}
-
-	var inPre = false, inBold = false;
-	for (var i = 0, len = source.length; i < len; i++) {
-		var ch = source.charAt(i);
-
-		if (ch === '\n') {
-			push(i, currentTagType());
-			r.push({
-				tagName: 'br'
-			});
-			lastPushedTo = i + 1;
-		} else if (ch === '`') {
-			push(i, currentTagType());
-			lastPushedTo = i + 1;
-			inPre = !inPre;
-		} else if (ch === '*') {
-			push(i, currentTagType());
-			lastPushedTo = i + 1;
-			inBold = !inBold;
-		}
-	}
-	push(source.length, currentTagType());
-
-	return r;
 }
 
 // https://docs.docker.com/reference/builder/
